# ApiSpec.yaml
openapi: '3.0.3'
info:
  description: APIs for managing recepies
  title:  RECIPIE API
  version: '3.0.0'
servers: [
  {
'url': 'http://localhost:8080/api/v1/recipies',
      'description': 'Base path of the API',

}]
...
security:
  - BearerAuth: []
paths: {
...
  '/getRecipies':{
    
        'post': {
        'operationId': 'getRecipies',
        'description': 'API to return recipies based on search criteria if no criteria is provided, it fetches all',
        'parameters': [
          {
            'name': 'traceId',
            'in': 'header',
            'required': true,
            'description': 'Trace id used to trace the request from beginning to the end, example - 12345678',
            'schema': {
              'type': 'string'
            }
          },
           {
            'name':'pageNo',
            'in': 'query',
            'description': 'fetch based on pageNo to support pagination',
            'schema': {
              'type': 'integer'
            }
          },
          {
            'name':'size',
            'in': 'query',
            'description': 'fetch based on size to support limiting the serach results',
            'schema': {
              'type': 'integer'
            }
          }
         
        ],
       'requestBody':
       {
          'required':false,
          'description': 'params required to search recipie/recipies',
            'content': {
              'application/json': {
                'schema': {
                  '$ref': '#/components/schemas/SearchResponseDTO'
                }
              }
            }
       },
        'responses': {
          '200': {
            'description': 'SuccessFul retrival of a recipe/recipies ',
            'content': {
              'application/json': {
                'schema': {
                  '$ref': '#/components/schemas/ResponseDTO'
                }
              }
            }
          },
          '400': {
                  '$ref': '#/components/responses/BadRequestException'
                },
        '401': {
                  '$ref': '#/components/responses/AuthenticationException'
                },
        '403': {
                  '$ref': '#/components/responses/AuthorizationException'
                },
        '404': {
                  '$ref': '#/components/responses/ResourceNotFoundException'
                },
                
        '415': {
                  '$ref': '#/components/responses/UnsupportedTypeException'
                },
        '500': {
                  '$ref': '#/components/responses/ServerException'
                }          
              
        }
          
        }
     
   
    
  },
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: jwt
  responses:
    BadRequestException:
       description: user input is invalid
       content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errors'
          example:
            error: Invalid input field  recipieName is required
    AuthenticationException:
      description: when an invalid credentials is supplied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errors'
          example:
            error: Invalid credentials supplied
    AuthorizationException:
      description: the user when doesnt have privillage
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errors'
          example:
            error: Access denied due to insufficient privilage
    ResourceNotFoundException:
      description: requested resource/item or not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errors'
          example:
            error: No such resource/item available 
    UnsupportedTypeException:
      description: requested body isnt the allowed format
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errors'
          example:
            error: request isnt of the type supported        
    ServerException:
      description: when there is an internal issues
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errors'
          example:
            error: Technical issues when operation is performed
  schemas: 
    SearchResponseDTO:
       type: object
       properties:
        recipies:
          description: 'List of recepies'
          type: array
          items:
            $ref: '#/components/schemas/ResponseDTO'
    ResponseDTO:
       type: object
       properties:
        receipeId:
          description: 'ID of the recepie'
          type: string
          example: 5672
        recipieName:
          description: 'Name of the recepie'
          type: string
          example: Dosa
        suitableFor:
          description: 'how many people can eat'
          type: integer
          example: 1
        instructions:
          description: 'how the recepie is made'
          type: string
          example: made with flour and water mixed
        veg:
          description: 'If the recepie is veg'
          type: boolean
          example: true
        ingredients:
          description: 'Content of the recepie'
          type: array
          items:
            $ref: '#/components/schemas/IngredientsDTO'
    IngredientsDTO:
       type: object
       properties:
        itemName:
          description: 'names of the ingredients'
          type: string
          example: 'tomato, rice, masala'
...
    Errors:
      type: object
      properties:
       status:
          description: 'status of the error code'
          type: integer
       message:
          description: 'error message'
          type: string
       timestamp:
          description: 'timestamp value during when exception occured'
          type: string
       path:
         description: 'resource path which resulted to exception'
         type: string
          
      example: 
        BadRequest:
          status: 400
          message: 'Bad request Invalid input field  recipieName is required'
          timestamp: '2020-11-29T09:42:12.287+00:00'
          path: "/addRecipie"
        Unauthorized:
          status: 401
          message: 'Invalid credentials supplied'
          timestamp: '2020-11-29T09:42:12.287+00:00'
          path: "/addRecipie"
    
        ServerError:
          status: 500
          message: 'Technical issues when operation is performed'
          timestamp: '2020-11-29T09:42:12.287+00:00'
          path: "/addRecipie"
        NotFound:
          status: 404
          message: 'Requested resource or item not found'
          timestamp: '2020-11-29T09:42:12.287+00:00'
          path: "/addRecipie"